Refactor files & tests
