+++ CHANGELOG.md
@@ -2,10 +2,11 @@
 
 All notable changes to this project will be documented in this file. This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [4.3.0](https://github.com/uttori/uttori-plugin-renderer-markdown-it/compare/v4.2.1...v4.3.0) - 2021-11-24
+## [4.3.0](https://github.com/uttori/uttori-plugin-renderer-markdown-it/compare/v4.2.1...v4.3.0) - 2021-11-26
 
 - 🧰 Support for `[^1] ... [^1]: Footnote` Footnotes based on [markdown-it-footnote](https://github.com/markdown-it/markdown-it-footnote)
 - 🛠 Fix Table of Contents HTML generation to be valid HTML (`UL` in `LI`)
+- 🛠 Refactor files to be more function specific
 - 🎁 Update dev dependencies
 
 ## [4.2.1](https://github.com/uttori/uttori-plugin-renderer-markdown-it/compare/v4.2.0...v4.2.1) - 2021-10-10
