+++ esm/index.js
@@ -6,3 +6,8 @@ export default MarkdownItRenderer;
 
 export { default as markdownItPlugin } from '../src/markdown-it-plugin';
 export { default as MarkdownItRenderer } from '../src/index.js';
+export { default as footnotes } from '../src/footnotes';
+export { default as toc } from '../src/toc';
+export { default as uttoriInline } from '../src/uttori-inline';
+export { default as wikilinks } from '../src/wikilinks';
+export { default as youtube } from '../src/youtube';
