+++ test/index.test.js
@@ -142,93 +142,3 @@ test('MarkdownItRenderer.render(content, config): replaces missing links with a 
   t.is(MarkdownItRenderer.render('[Test]()'), '<p><a href="/test">Test</a></p>');
   t.is(MarkdownItRenderer.render('[CrAzY CaSe SpAcEd]()'), '<p><a href="/crazy-case-spaced">CrAzY CaSe SpAcEd</a></p>');
 });
-
-test('MarkdownItRenderer.render(content, config): prepends the baseURL when set', (t) => {
-  t.is(MarkdownItRenderer.render('[Test]', { uttori: { baseUrl: '/wiki' } }), '<p>[Test]</p>');
-  t.is(MarkdownItRenderer.render('[Test]()', { uttori: { baseUrl: '/wiki' } }), '<p><a href="/wiki/test">Test</a></p>');
-  t.is(MarkdownItRenderer.render('[CrAzY CaSe SpAcEd]()', { uttori: { baseUrl: '/wiki' } }), '<p><a href="/wiki/crazy-case-spaced">CrAzY CaSe SpAcEd</a></p>');
-});
-
-test('MarkdownItRenderer.render(content, config): protects SEO by ignoring unknown domains', (t) => {
-  t.is(MarkdownItRenderer.render('[Test]', { uttori: { allowedExternalDomains: ['example.org'] } }), '<p>[Test]</p>');
-  t.is(MarkdownItRenderer.render('[Test]()', { uttori: { allowedExternalDomains: ['example.org'] } }), '<p><a href="/test">Test</a></p>');
-  t.is(MarkdownItRenderer.render('[Test](http://example.org/wiki/test)', { uttori: { allowedExternalDomains: ['example.org'] } }), '<p><a href="http://example.org/wiki/test" rel="external noopener noreferrer">Test</a></p>');
-  t.is(MarkdownItRenderer.render('[Test](https://example.org/wiki/test)', { uttori: { allowedExternalDomains: ['example.org'] } }), '<p><a href="https://example.org/wiki/test" rel="external noopener noreferrer">Test</a></p>');
-  t.is(MarkdownItRenderer.render('[Test](http://evil.org/wiki/test)', { uttori: { allowedExternalDomains: ['example.org'] } }), '<p><a href="http://evil.org/wiki/test" rel="external nofollow noopener noreferrer">Test</a></p>');
-  t.is(MarkdownItRenderer.render('[Test](https://evil.org/wiki/test)', { uttori: { allowedExternalDomains: ['example.org'] } }), '<p><a href="https://evil.org/wiki/test" rel="external nofollow noopener noreferrer">Test</a></p>');
-});
-
-test('MarkdownItRenderer.render(content, config): can set external links to open in a new window', (t) => {
-  t.is(MarkdownItRenderer.render('[Test]', { uttori: { openNewWindow: true, allowedExternalDomains: ['example.org'] } }), '<p>[Test]</p>');
-  t.is(MarkdownItRenderer.render('[Test]()', { uttori: { openNewWindow: true, allowedExternalDomains: ['example.org'] } }), '<p><a href="/test">Test</a></p>');
-  t.is(MarkdownItRenderer.render('[Test](/wiki/test)', { uttori: { openNewWindow: true, allowedExternalDomains: ['example.org'] } }), '<p><a href="/wiki/test">Test</a></p>');
-  t.is(MarkdownItRenderer.render('[Test](/wiki/test)', { uttori: { openNewWindow: true, allowedExternalDomains: ['example.org'] } }), '<p><a href="/wiki/test">Test</a></p>');
-  t.is(MarkdownItRenderer.render('[Test](http://example.org/wiki/test)', { uttori: { openNewWindow: true, allowedExternalDomains: ['example.org'] } }), '<p><a href="http://example.org/wiki/test" rel="external noopener noreferrer" target="_blank">Test</a></p>');
-  t.is(MarkdownItRenderer.render('[Test](https://example.org/wiki/test)', { uttori: { openNewWindow: true, allowedExternalDomains: ['example.org'] } }), '<p><a href="https://example.org/wiki/test" rel="external noopener noreferrer" target="_blank">Test</a></p>');
-  t.is(MarkdownItRenderer.render('[Test](http://evil.org/wiki/test)', { uttori: { openNewWindow: true, allowedExternalDomains: ['example.org'] } }), '<p><a href="http://evil.org/wiki/test" rel="external nofollow noopener noreferrer" target="_blank">Test</a></p>');
-  t.is(MarkdownItRenderer.render('[Test](https://evil.org/wiki/test)', { uttori: { openNewWindow: true, allowedExternalDomains: ['example.org'] } }), '<p><a href="https://evil.org/wiki/test" rel="external nofollow noopener noreferrer" target="_blank">Test</a></p>');
-});
-
-test('MarkdownItRenderer.render(content, config): can render a WikiLink', (t) => {
-  const markdown = 'A deep [[Link]]';
-  const output = '<p>A deep <a href="link">Link</a></p>';
-  t.is(MarkdownItRenderer.render(markdown, MarkdownItRenderer.defaultConfig()), output);
-});
-
-test('MarkdownItRenderer.render(content, config): can render multiple WikiLinks', (t) => {
-  const markdown = 'A deep [[Link]] and a deeper [[hole]] but maybe a [[lake|big lake]]';
-  const output = '<p>A deep <a href="link">Link</a> and a deeper <a href="hole">hole</a> but maybe a <a href="lake">big lake</a></p>';
-  t.is(MarkdownItRenderer.render(markdown, MarkdownItRenderer.defaultConfig()), output);
-});
-
-test('MarkdownItRenderer.render(content, config): can render a WikiLink with custom anchor text', (t) => {
-  const markdown = 'A deep [[LINK|hole]]';
-  const output = '<p>A deep <a href="link">hole</a></p>';
-  t.is(MarkdownItRenderer.render(markdown, MarkdownItRenderer.defaultConfig()), output);
-});
-
-test('MarkdownItRenderer.render(content, config): can render a WikiLink with a bracket character in the text', (t) => {
-  // eslint-disable-next-line no-useless-escape
-  // eslint-disable-next-line quotes
-  const markdown = `A deep [[LINK|\\[hole]]`;
-  const output = '<p>A deep <a href="link">\\[hole</a></p>';
-  t.is(MarkdownItRenderer.render(markdown, MarkdownItRenderer.defaultConfig()), output);
-});
-
-test('MarkdownItRenderer.render(content, config): can render a WikiLink with a bad text part', (t) => {
-  const markdown = 'A deep [[LINK|]]';
-  const output = '<p>A deep <a href="link">LINK</a></p>';
-  t.is(MarkdownItRenderer.render(markdown, MarkdownItRenderer.defaultConfig()), output);
-});
-
-test('MarkdownItRenderer.render(content, config): does not error with a WikiLink with another WikiLink starting before the first closed', (t) => {
-  const markdown = 'A deep [[LINK[[]]';
-  const output = '<p>A deep <a href=""></a>[[LINK<a href=""></a></p>';
-  t.is(MarkdownItRenderer.render(markdown, MarkdownItRenderer.defaultConfig()), output);
-});
-
-test('MarkdownItRenderer.render(content, config): does not error with a WikiLink with a line break in the text part', (t) => {
-  const markdown = `A deep [[LINK|
-]]`;
-  const output = `<p>A deep [[LINK|
-]]</p>`;
-  t.is(MarkdownItRenderer.render(markdown, MarkdownItRenderer.defaultConfig()), output);
-});
-
-test('MarkdownItRenderer.render(content, config): can render a WikiLink with a baseUrl', (t) => {
-  const markdown = 'A deep [[Link]]';
-  const output = '<p>A deep <a href="/wiki/link">Link</a></p>';
-  t.is(MarkdownItRenderer.render(markdown, { ...MarkdownItRenderer.defaultConfig(), uttori: { ...MarkdownItRenderer.defaultConfig().uttori, baseUrl: '/wiki' } }), output);
-});
-
-test('MarkdownItRenderer.render(content, config): can render an Image with lazy loading', (t) => {
-  const markdown = '![](example.png "image title")';
-  const output = '<p><img src="example.png" alt="" title="image title" loading="lazy"></p>';
-  t.is(MarkdownItRenderer.render(markdown, { ...MarkdownItRenderer.defaultConfig(), uttori: { ...MarkdownItRenderer.defaultConfig().uttori, lazyImages: true } }), output);
-});
-
-test('MarkdownItRenderer.render(content, config): can render an Image without lazy loading', (t) => {
-  const markdown = '![](example.png "image title")';
-  const output = '<p><img src="example.png" alt="" title="image title"></p>';
-  t.is(MarkdownItRenderer.render(markdown, { ...MarkdownItRenderer.defaultConfig(), uttori: { ...MarkdownItRenderer.defaultConfig().uttori, lazyImages: false } }), output);
-});
